[archiso/build] Fix `cleanup` to propagate exit code
This change fixes the prior implementation which assumed that an exit
code would have to be set in a `trap` which is incorrect (the exit code
is preserved on an `EXIT` trap). Furthermore the exit code that was
attempted to be captured was of the local variable being set, not the
previous called command, meaning that `$e` was always `0`. Whoops...

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>